DownloadSingleDataStructure (#85)
* DownloadSingleDataStructure

* Expand to match data structure(s) and ignore versions

Being able to select which data structures you care about has been requested a
few times, and we've discussed it internally a bit — so thanks for getting the
ball rolling with this PR :)

The download command was originally implemented to help keep your local data
structures in sync with whatever is currently deployed in your BDP console's
dev environment. Allowing selection of specific versions as part of that might
be a bit confusing in that context.

I hope the changes I’ve made here don’t go against what you were aiming to
achieve. If you're happy with them, I think we’re good to get this merged in.

If your main goal was actually to be able to pull down specific versions, I’d
be really interested to hear more about your use case. We might want to support
that in a different way — perhaps a new command for schema browsing, like:

`snowplow-cli ds show iglu:x/x/x/1-0-0`

---------

Co-authored-by: james <james@snowplowanalytics.com>